added cover/avater placeholder to space page
