+++ apps/web/app/space/(entity)/[id]/[entityId]/layout.tsx
@@ -127,7 +127,7 @@ export default async function ProfileLayout(props: Props) {
         <EntityPageCover avatarUrl={profile.avatarUrl} coverUrl={profile.coverUrl} />
         <EntityPageContentContainer>
           <div className="space-y-2">
-            <EditableHeading spaceId={params.id} entityId={entityId} />
+            <EditableHeading spaceId={params.id} entityName={profile.name ?? ''} entityId={entityId} />
             <EntityPageMetadataHeader id={profile.id} entityName={profile.name ?? ''} spaceId={params.id} />
           </div>
